backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
numpy==1.26.4
scipy==1.14.1
jinja2==3.1.4
python-multipart==0.0.12
paho-mqtt==1.6.1
